chore: Update phone and isBlocked variables in UsersTableEditor
Update the `phone` variable to use the value at column index 2 instead of 1, and update the `isBlocked` variable to use the value at column index 6 instead of 4 in the `getTableCellEditorComponent` method of the `UsersTableEditor` class.

Update the selected button and change the frame to the second one in the `MenuAdmin` class.

Update the `pwd` field to display the password for both admin and employee accounts in the `EditAccount` class.